total comp bug fix
bug fix for case when user re-navigates to previously opened profile and some of the fields would appear blank
